test(sandbox): move topic-isolation test to e2e + document test layout
The sandbox topic-isolation test opens AsyncSessions and runs DB writes,
so it belongs in tests/e2e/ not tests/integration/. Placing it under
integration ran it during every pre-push CI sweep and made the suite
tip over the moment a new migration landed without re-`upgrade head`ing
the per-slot test DB.

- Move tests/integration/sandbox/test_topic_isolation.py to
  tests/e2e/test_sandbox_topic_isolation.py, inlining the previously
  shared fixtures (the e2e conftest's fake_opensandbox doesn't echo
  sandbox_id, which the reuse-path assertions rely on).
- Drop the now-empty tests/integration/sandbox/ tree.
- Document the rule in AGENTS.md so future work picks the right
  directory: anything touching Postgres / Redis / OpenSandbox / FastAPI
  app goes in tests/e2e/.
